Make renderTextEdit a Part as well
